+++ examples/table/mui/dataGridPro/src/App.tsx
@@ -36,6 +36,7 @@ const App: React.FC = () => {
                             list: PostsList,
                         },
                     ]}
+                    disableTelemetry={true}
                 />
             </RefineSnackbarProvider>
         </ThemeProvider>