+++ packages/mui/src/components/layout/sider/index.spec.tsx
@@ -67,7 +67,20 @@ describe("Sider", () => {
     it("should render only allowed menu items", async () => {
         const { getAllByText, queryByText } = render(<Sider />, {
             wrapper: TestWrapper({
-                resources: [{ name: "posts" }, { name: "users" }],
+                resources: [
+                    {
+                        name: "posts",
+                        list: function list() {
+                            return <div>render me!</div>;
+                        },
+                    },
+                    {
+                        name: "users",
+                        list: function list() {
+                            return <div>render me!</div>;
+                        },
+                    },
+                ],
                 accessControlProvider: {
                     can: ({ action, resource }) => {
                         if (action === "list" && resource === "posts") {